streamlit_app/requirements.txt
asyncio
streamlit-webrtc
streamlit
firebase-admin
opencv-python
pyrebase
pillow
PyJWT
tensorflow
numpy